claude-wrapper/requirements.txt
bubbletea-chat[llm]
python-dotenv
anthropic
